Refactor: Flatten modules and consolidate semantic model
Flatten `src/errors/` into `src/errors.rs` and merge `src/semantic/assembly_model.rs` into `src/semantic/assembler.rs`. This reduces module depth and improves cohesion by keeping data structures with their producer logic. No logic changes.
