+++ src/semantic/conversion.rs
@@ -42,7 +42,7 @@ use super::patterns::detect_iterator_pattern;
 use crate::ast::Expr;
 use crate::errors::GlossaError;
 use crate::morphology::{self};
-use crate::semantic::assembly_model::AssembledStatement;
+use crate::semantic::assembler::AssembledStatement;
 use crate::semantic::model::{AnalyzedExpr, AnalyzedExprKind, AnalyzedStatement};
 use crate::semantic::resolver::Scope;
 use crate::semantic::types::GlossaType;
