+++ src/semantic/expressions.rs
@@ -16,7 +16,7 @@ use crate::ast::{Expr, Statement};
 use crate::errors::GlossaError;
 use crate::morphology::{self, DisambiguationContext, analyze_article, disambiguate, resolve_best};
 use crate::semantic::assembler::Assembler;
-use crate::semantic::assembly_model::Literal;
+use crate::semantic::assembler::Literal;
 use crate::semantic::model::{AnalyzedExpr, AnalyzedExprKind};
 use crate::semantic::resolver::Scope;
 use crate::semantic::types::GlossaType;
