+++ src/semantic/mod.rs
@@ -34,7 +34,6 @@
 pub mod assembler;
 #[cfg(test)]
 mod assembler_tests;
-pub(crate) mod assembly_model;
 #[cfg(test)]
 mod classification_tests;
 pub(crate) mod control_flow;
@@ -50,8 +49,7 @@ mod types;
 
 pub use crate::morphology::{DisambiguationContext, analyze_article, disambiguate, resolve_best};
 pub(crate) use assembler::Assembler;
-pub use assembler::AssemblyError;
-pub use assembly_model::{AssembledStatement, Constituent, Literal};
+pub use assembler::{AssembledStatement, AssemblyError, Constituent, Literal};
 pub use model::*;
 pub use resolver::*;
 pub use types::*;
